WineANNPrediction/requirements.txt
Flask
pandas
scikit-learn
numpy
gunicorn
imblearn
tensorflow-cpu